projects/FORT-contracts/requirements.txt
fastapi
uvicorn[standard]
pynacl
python-dotenv
